fix: fixed screenshot client logic to handle missing configuration metadata

+++ src/screenshot.py
@@ -3,38 +3,41 @@ from pathlib import Path
 
 from playwright.sync_api import Page
 
-from notify import Notifier
+from typing import Type
 from src import logging
 
 
 class Screenshot:
-    def __init__(self, notifier: Notifier, path: Path):
+    def __init__(self, notifier: Type["Notifier"], path: Path):
         self.notifier = notifier
         self.path = path
 
         self.screenshot_path = None
+        self.disabled = self.notifier.disabled
 
     def take_screenshot(
         self, page: Page, filename: str = "screenshot", full_page: bool = False
     ):
-        logging.debug(" ### Taking screenshot ###")
-        self.screenshot_path = self.path / f"{filename}.png"
+        if not self.disabled:
+            logging.debug(" ### Taking screenshot ###")
+            self.screenshot_path = self.path / f"{filename}.png"
 
-        page.screenshot(path=self.screenshot_path, full_page=full_page)
-        self.notifier.send_image(image_location=self.screenshot_path)
+            page.screenshot(path=self.screenshot_path, full_page=full_page)
+            self.notifier.send_image(image_location=self.screenshot_path)
 
     def remove_screenshot(self):
-        try:
-            if self.screenshot_path:
-                os.remove(self.screenshot_path)
-            else:
+        if not self.disabled:
+            try:
+                if self.screenshot_path:
+                    os.remove(self.screenshot_path)
+                else:
+                    raise ScreenshotNotTaken(
+                        "Screenshot has not been taken yet, thus cannot be removed!"
+                    )
+            except Exception:
                 raise ScreenshotNotTaken(
                     "Screenshot has not been taken yet, thus cannot be removed!"
                 )
-        except Exception:
-            raise ScreenshotNotTaken(
-                "Screenshot has not been taken yet, thus cannot be removed!"
-            )
 
 
 class ScreenshotNotTaken(Exception):
